requirements.txt
fastapi
uvicorn
motor
python-jose
Crypto
python-dotenv
python-multipart
passlib[bcrypt]
PyJWT
pydantic
